chore(common): Remove PropertyMetadata
BREAKING CHANGE: Remove PropertyMetadata. Use JsonEntityStore instead
