+++ packages/platform/common/src/domain/index.ts
@@ -4,4 +4,3 @@ export * from "./PlatformRouteDetails";
 export * from "./ControllerProvider";
 export * from "./EndpointMetadata";
 export * from "./HandlerMetadata";
-export * from "./PropertyMetadata";
